No need for a new viewController to only mange the countDown, it will be only one ViewController wit different states
